Refactored DispatchProcess.java
Now sqatt compiles all java files in a directory
